+++ WPF.Utils/InvokeTextBox.cs
@@ -11,11 +11,13 @@ namespace WPF.Utils
     {
         public static void Clear(TextBox control)
         {
+            if (control is null)
+                return;
             void Work(TextBox inControl)
             {
                 inControl.Clear();
             }
-            if (!(control is null) && !control.CheckAccess())
+            if (!control.CheckAccess())
                 if (!(Application.Current is null))
                     Application.Current.Dispatcher.Invoke(DispatcherPriority.Normal, new Action(() =>
                     {
@@ -32,11 +34,13 @@ namespace WPF.Utils
 
         public static void SetText(TextBox control, string text)
         {
+            if (control is null)
+                return;
             void Work(TextBox inControl, string inText)
             {
                 inControl.Text = inText;
             }
-            if (!(control is null) && !control.CheckAccess())
+            if (!control.CheckAccess())
                 if (!(Application.Current is null))
                     Application.Current.Dispatcher.Invoke(DispatcherPriority.Normal, new Action(() =>
                     {
@@ -54,11 +58,13 @@ namespace WPF.Utils
         public static string GetText(TextBox control)
         {
             var result = string.Empty;
+            if (control is null)
+                return result;
             string Work(TextBox inControl)
             {
                 return inControl.Text;
             }
-            if (!(control is null) && !control.CheckAccess())
+            if (!control.CheckAccess())
                 if (!(Application.Current is null))
                     Application.Current.Dispatcher.Invoke(DispatcherPriority.Normal, new Action(() =>
                     {
@@ -76,11 +82,13 @@ namespace WPF.Utils
 
         public static void AddText(TextBox control, string text)
         {
+            if (control is null)
+                return;
             void Work(TextBox inControl, string inText)
             {
                 inControl.Text += inText + Environment.NewLine;
             }
-            if (!(control is null) && !control.CheckAccess())
+            if (!control.CheckAccess())
                 if (!(Application.Current is null))
                     Application.Current.Dispatcher.Invoke(DispatcherPriority.Normal, new Action(() =>
                     {
@@ -97,11 +105,13 @@ namespace WPF.Utils
 
         public static void AddTextFormat(TextBox control, Stopwatch sw, string text)
         {
+            if (control is null)
+                return;
             void Work(TextBox inControl, Stopwatch inSw, string inText)
             {
                 inControl.Text += $@"[{inSw.Elapsed}] {inText}" + Environment.NewLine;
             }
-            if (!(control is null) && !control.CheckAccess())
+            if (!control.CheckAccess())
                 if (!(Application.Current is null))
                     Application.Current.Dispatcher.Invoke(DispatcherPriority.Normal, new Action(() =>
                     {
@@ -114,6 +124,34 @@ namespace WPF.Utils
                     }));
             else
                 Work(control, sw, text);
+        }
+
+        public static void AddTextFormat(TextBox control, DateTime dt, string text)
+        {
+            if (control is null)
+                return;
+            void Work(TextBox inControl, DateTime inDt, string inText)
+            {
+                inControl.Text += $@"[{inDt}] {inText}" + Environment.NewLine;
+            }
+            if (!control.CheckAccess())
+                if (!(Application.Current is null))
+                    Application.Current.Dispatcher.Invoke(DispatcherPriority.Normal, new Action(() =>
+                    {
+                        Work(control, dt, text);
+                    }));
+                else
+                    control.Dispatcher.Invoke(DispatcherPriority.Normal, new Action(() =>
+                    {
+                        Work(control, dt, text);
+                    }));
+            else
+                Work(control, dt, text);
+        }
+
+        public static void AddTextFormat(TextBox control, string text)
+        {
+            AddTextFormat(control, DateTime.Now, text);
         }
     }
 }